CURA-5164 Fix code style in PluginRegistry

+++ UM/PluginRegistry.py
@@ -73,16 +73,14 @@ class PluginRegistry(QObject):
         self._plugin_locations = []  # type: List[str]
         self._folder_cache = {}      # type: Dict[str, List[Tuple[str, str]]]
 
-        self._supported_file_types = {"umplugin": "Uranium Plugin"}
-
-        # File to store plugin info, such as which ones to install/remove and which ones are disabled.
-        # Cannot load this here because we don't know the actual Application name yet, so it will result in incorrect
-        # directory name if we try to get it from Resources now.
-        self._plugin_config_filename = None #type: Optional[str]
+        self._supported_file_types = {"umplugin": "Uranium Plugin"} # type: Dict[str, str]
 
     def initializeBeforePluginsAreLoaded(self) -> None:
         config_path = Resources.getConfigStoragePath()
-        self._plugin_config_filename = os.path.join(os.path.abspath(config_path), "plugins.json")
+
+        # File to store plugin info, such as which ones to install/remove and which ones are disabled.
+        # At this point we can load this here because we already know the actual Application name, so the directory name
+        self._plugin_config_filename = os.path.join(os.path.abspath(config_path), "plugins.json") # type: str
 
         from UM.Settings.ContainerRegistry import ContainerRegistry
         container_registry = ContainerRegistry.getInstance()
@@ -139,12 +137,13 @@ class PluginRegistry(QObject):
         from UM.Settings.ContainerRegistry import ContainerRegistry
         container_registry = ContainerRegistry.getInstance()
         with container_registry.lockFile():
-            with open(self._plugin_config_filename, "w", encoding = "utf-8") as f:
-                data = json.dumps({"disabled": self._disabled_plugins,
-                                   "to_install": self._plugins_to_install,
-                                   "to_remove": self._plugins_to_remove,
-                                   })
-                f.write(data)
+            if os.path.exists(self._plugin_config_filename):
+                with open(self._plugin_config_filename, "w", encoding = "utf-8") as f:
+                    data = json.dumps({"disabled": self._disabled_plugins,
+                                       "to_install": self._plugins_to_install,
+                                       "to_remove": self._plugins_to_remove,
+                                       })
+                    f.write(data)
 
     # TODO:
     # - [ ] Improve how metadata is stored. It should not be in the 'plugin' prop
@@ -535,7 +534,7 @@ class PluginRegistry(QObject):
     ##  Try to find a module implementing a plugin
     #   \param plugin_id The name of the plugin to find
     #   \returns module if it was found None otherwise
-    def _findPlugin(self, plugin_id: str) -> types.ModuleType:
+    def _findPlugin(self, plugin_id: str) -> Optional[types.ModuleType]:
         location = None
         for folder in self._plugin_locations:
             location = self._locatePlugin(plugin_id, folder)
@@ -579,9 +578,9 @@ class PluginRegistry(QObject):
             if file == plugin_id and os.path.exists(os.path.join(file_path, "__init__.py")):
                 return folder
             else:
-                file_path = self._locatePlugin(plugin_id, file_path)
-                if file_path:
-                    return file_path
+                plugin_path = self._locatePlugin(plugin_id, file_path)
+                if plugin_path:
+                    return plugin_path
 
         return None
 
@@ -729,7 +728,7 @@ class PluginRegistry(QObject):
     #   \return \type{string} The absolute path to the plugin or an empty string if the plugin could not be found.
     def getPluginPath(self, plugin_id: str) -> Optional[str]:
         if plugin_id in self._plugins:
-            plugin = self._plugins[plugin_id]
+            plugin = self._plugins.get(plugin_id)
         else:
             plugin = self._findPlugin(plugin_id)
